pyproject.toml
[project]
name = "interview_prep"
version = "0.1.0"
description = "interview-prep using crewAI"
authors = [{ name = "Veronica Schembri", email = "veronicaschembri@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.118.0,<1.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
interview_prep = "interview_prep.main:run"
run_crew = "interview_prep.main:run"
demo = "interview_prep.demo:run_demo"
train = "interview_prep.main:train"
replay = "interview_prep.main:replay"
test = "interview_prep.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
